% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf-compare-sim.R
\name{abf.fn}
\alias{abf.fn}
\title{Joint Approximate Bayes' factors (ABFs) for two case-control studies with shared controls}
\usage{
abf.fn(sim, msnps, mT1, mT2)
}
\arguments{
\item{sim}{output from phen.gen.fn}

\item{msnps}{vector of SNPs to consider in models}

\item{mT1}{maximum number of causal variants for trait 1}

\item{mT2}{maximum number of causal variants for trait 2}
}
\value{
data.frame of joint model ABFs (each row is a model): column 1 has joint ABFs, remaining columns specify inclusion/excusion of SNPs in each joint model
add column of 1's to the models for the trait2*effect variable
}
\description{
Joint Approximate Bayes' factors (ABFs) for two case-control studies with shared controls
}
\author{
Jenn Asimit
}
