% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf-compare-sim.R
\name{abfcalc.format}
\alias{abfcalc.format}
\title{Approximate Bayes' factors (ABFs) for one case-control study}
\usage{
abfcalc.format(sim)
}
\arguments{
\item{sim}{output from phen.gen.fn}
}
\value{
list of data.frames for each trait; column 1 gives case-control status (1/0) and remaining columns form a SnpMatrix object
}
\description{
Approximate Bayes' factors (ABFs) for one case-control study
}
