% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf-compare-sim.R
\name{abfT1.fn}
\alias{abfT1.fn}
\title{Approximate Bayes' factors (ABFs) for one case-control study}
\usage{
abfT1.fn(data1, mT1 = 3, msnps)
}
\arguments{
\item{data1}{data.frame that has case-control status (1-0) in column 1 and remaining columns are the genotype scores (rows are individuals)}

\item{mT1}{maximum number of causal variants}

\item{msnps}{vector of SNPs to consider in models}
}
\value{
data.frame of model ABFs (each row is a model): column 1 has ABFs, remaining columns specify inclusion/excusion of SNPs in each model
}
\description{
Approximate Bayes' factors (ABFs) for one case-control study
}
\author{
Jenn Asimit
}
