% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf-compare-sim.R
\name{bf.compare.fn}
\alias{bf.compare.fn}
\title{Joint Approximate Bayes' factors (ABFs) for two case-control studies with shared controls}
\usage{
bf.compare.fn(sim, msnps, mT1, mT2)
}
\arguments{
\item{sim}{output from phen.gen.fn}

\item{msnps}{vector of SNPs to consider in models}

\item{mT1}{maximum number of causal variants for trait 1}

\item{mT2}{maximum number of causal variants for trait 2}
}
\value{
data.frame of multinomial logABFs and logistic logABFs
find marginal ABFs for each trait
find joint ABFs for traits 1 and 2
}
\description{
Joint Approximate Bayes' factors (ABFs) for two case-control studies with shared controls
}
\author{
Jenn Asimit
}
