% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf-compare-sim.R
\name{bf.relations.fn}
\alias{bf.relations.fn}
\title{Relationship between multinomial ABFs and logistic ABFs}
\usage{
bf.relations.fn(BFs, Bplot = FALSE)
}
\arguments{
\item{BFs}{data.frame of multinomial logABFs BF12 and logistic logABFs}

\item{Bplot}{logical, if TRUE plot ABF12 against (log(ABF1)+log(ABF2)}
}
\value{
fitted regression model R2 and coefficient estimate summaries from log(jointABF) ~ (log(ABF1)+log(ABF2))
}
\description{
Relationship between multinomial ABFs and logistic ABFs
}
