% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{gen.fn}
\alias{gen.fn}
\title{Convert genotype calls at a SNP to genotype score}
\usage{
gen.fn(i, gcalls)
}
\arguments{
\item{i}{SNP index}

\item{gcalls}{data.frame of genotype calls where rows are SNPs and columns are individuals}
}
\value{
genotype score of SNP
}
\description{
Convert genotype calls at a SNP to genotype score
}
