% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf-compare-sim.R
\name{phen.gen.fn}
\alias{phen.gen.fn}
\title{Generate case-control data with shared controls for traits 1 and 2}
\usage{
phen.gen.fn(beta1 = c(-2.3, 0.2, 0.2), beta2 = c(-2.3, 0.2, 0.2), snpG,
  N0 = 100, N1 = 100, N2 = 100, causals1.ind, causals2.ind)
}
\arguments{
\item{beta1}{vector of model parameters for trait 1: (log(prevalence),causal variant effects)}

\item{beta2}{vector of model parameters for trait 2: (log(prevalence),causal variant effects)}

\item{snpG}{matrix of genotype scores for null data (rows=SNPs, columns=individuals)}

\item{N0}{number of shared controls}

\item{N1}{number of cases for trait 1}

\item{N2}{number of cases for trait 2}

\item{causals1.ind}{indices of trait 1 causal variants with respect to snpG}

\item{causals2.ind}{indices of trait 2 causal variants with respect to snpG}
}
\value{
list consisting of G=genotype matrix (rows=indiv, cols=snps), y=vector of case-control status (0=control,1=trait 1 case, 2=trait 2 case)
beta\link{1} = log(prev)
}
