% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf-compare-sim.R
\name{T1modsmat.fn}
\alias{T1modsmat.fn}
\title{Generate model matrix of all possible models with maximum specified size}
\usage{
T1modsmat.fn(mT1, msnps)
}
\arguments{
\item{mT1}{maximum model size}

\item{msnps}{SNPs for model matrix}
}
\value{
model matrix for SNPs specified in msnps; columns are SNPs, each row corresponds to a model
}
